Fix empty DNS Name Server option
Fix a bug where, if the DNS Name Servers option of a subnet
has become empty, no attempt is made to update the equivalent
option in Infoblox.

This is an erorr since the DNS Name Servers
option may previously have valid entries, and therefore those
entries need to be removed from Infoblox.

Change-Id: I9c246295b02cac58c9d19e104765063416dc7deb